+++ react/components/Autocomplete/components/CustomListItem/CustomListItem.tsx
@@ -6,6 +6,7 @@ import ProductPrice from "vtex.store-components/ProductPrice";
 
 interface CustomListItemProps {
   product: any;
+  onClick: () => void;
 }
 
 export class CustomListItem extends React.Component<CustomListItemProps> {
@@ -23,6 +24,7 @@ export class CustomListItem extends React.Component<CustomListItemProps> {
           }}
           page="store.product"
           className="no-underline"
+          onClick={this.props.onClick}
         >
           <article
             className={`${styles.element} flex flex-row justify-start items-center pa3 bg-animate hover-bg-light-gray`}